+++ src/AppBundle/Entity/User.php
@@ -7,6 +7,7 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\Collections\Criteria;
 use Doctrine\Common\Collections\Selectable;
 use Doctrine\ORM\Mapping as ORM;
+use Scheb\TwoFactorBundle\Model\Email\TwoFactorInterface;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Validator\Constraints as Assert;
@@ -17,7 +18,7 @@ use Symfony\Component\Validator\Constraints as Assert;
  *
  * @UniqueEntity("canonicalUsername", errorPath="username")
  */
-class User implements UserInterface {
+class User implements UserInterface, TwoFactorInterface {
     /**
      * @ORM\Column(type="bigint")
      * @ORM\GeneratedValue(strategy="AUTO")
@@ -133,6 +134,24 @@ class User implements UserInterface {
      * @var bool
      */
     private $nightMode = false;
+
+    /**
+     * @ORM\Column(type="boolean", options={"default": false})
+     *
+     * @Assert\Expression("value === false || this.getEmail() !== null",
+     *     message="Two-factor cannot be enabled without providing an email address.",
+     *     groups={"editing"})
+     *
+     * @var bool
+     */
+    private $twoFactorEnabled = false;
+
+    /**
+     * @ORM\Column(type="integer", nullable=true)
+     *
+     * @var int|null
+     */
+    private $emailAuthCode;
 
     public function __construct() {
         $this->created = new \DateTime('@'.time());
@@ -382,6 +401,41 @@ class User implements UserInterface {
      */
     public function setNightMode(bool $nightMode) {
         $this->nightMode = $nightMode;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isTwoFactorEnabled(): bool {
+        return $this->twoFactorEnabled;
+    }
+
+    /**
+     * @param bool $twoFactorEnabled
+     */
+    public function setTwoFactorEnabled(bool $twoFactorEnabled) {
+        $this->twoFactorEnabled = $twoFactorEnabled;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isEmailAuthEnabled() {
+        return $this->email !== null && $this->twoFactorEnabled;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getEmailAuthCode() {
+        return $this->emailAuthCode;
+    }
+
+    /**
+     * @param int|null $authCode
+     */
+    public function setEmailAuthCode($authCode) {
+        $this->emailAuthCode = $authCode;
     }
 
     /**