+++ src/AppBundle/Form/UserType.php
@@ -6,6 +6,7 @@ use Gregwar\CaptchaBundle\Type\CaptchaType;
 use Raddit\AppBundle\Entity\User;
 use Raddit\AppBundle\Form\EventListener\PasswordEncodingSubscriber;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
@@ -52,6 +53,11 @@ final class UserType extends AbstractType {
             $builder->add('verification', CaptchaType::class, [
                 'as_url' => true,
                 'reload' => true,
+            ]);
+        } else {
+            $builder->add('twoFactorEnabled', CheckboxType::class, [
+                'label' => 'user_form.two_factor_enabled',
+                'required' => false,
             ]);
         }